simplify simplex driver
